% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{apply_gsub}
\alias{apply_gsub}
\title{apply_gsub}
\usage{
apply_gsub(x, ...)
}
\arguments{
\item{x}{vector of strings to modify}

\item{...}{list syntax, replacement = c("words", "to", "replace")}
}
\description{
Batch together gsub commands
}
