% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tldr.R
\name{brief}
\alias{brief}
\title{brief}
\usage{
brief(x, n = 3)
}
\arguments{
\item{x}{character including all text}

\item{n}{number of sentences desired}
}
\description{
Summarises text based on desired number of sentences
}
