% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{c_sentence}
\alias{c_sentence}
\title{c_sentence}
\usage{
c_sentence(x, and = "and")
}
\arguments{
\item{x}{a vector}

\item{and}{a character string. The joining word}
}
\description{
concatenates words with commas. Replaces the last comma with an and.
}
