% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_resids}
\alias{check_resids}
\title{check_resids}
\usage{
check_resids(csq)
}
\arguments{
\item{csq}{htest object
importFrom ggplot2 ggplot aes_string geom_tile scale_fill_gradient2 theme_bw labs}
}
\description{
view residuals from an htest object
}
