% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{clippy}
\alias{clippy}
\title{clippy}
\usage{
clippy(x, row.names = F)
}
\arguments{
\item{x}{the data.frame}

\item{row.names}{a bool. TRUE preserves rownames}
}
\description{
sends a table to the clipboard so it can be pasted into excel - a wrapper for write.table
}
