% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_paper.R
\name{discussion}
\alias{discussion}
\title{discussion}
\usage{
discussion(
  key_finding,
  other_findings,
  context,
  limitations,
  whats_next,
  so_what,
  conclusion
)
}
\arguments{
\item{other_findings}{Secondary findings}

\item{context}{Compare to other findings}

\item{limitations}{What prevents firms conclusions being made?}

\item{so_what}{Why should anyone care about the findings?}

\item{conclusion}{restate main findings, final take-home message}

\item{key_findings}{Main outcome of the study}
}
\description{
Scaffold a discussion section
}
