% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame_assist.R
\name{find_vars}
\alias{find_vars}
\title{find_vars}
\usage{
find_vars(data, x, ...)
}
\arguments{
\item{data}{data.frame}

\item{x}{regex search string}

\item{...}{other arguments to regex}
}
\description{
Find vars in a dataframe
}
