% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plot.R
\name{frequency_plot}
\alias{frequency_plot}
\title{frequency_plot}
\usage{
frequency_plot(x, limit = 10)
}
\arguments{
\item{x}{data.frame containing only factors with identical levels}

\item{limit}{numeric. At what threshold should the percent label not be shown}
}
\description{
Visualise the percentage of responses to likert questions
}
