% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{glue_bracket}
\alias{glue_bracket}
\title{glue_bracket}
\usage{
glue_bracket(x, ..., round = 2, brackets = c("(", ")"), collapse = ", ")
}
\arguments{
\item{x}{a character or numeric}

\item{...}{a function. If provided will transform effects and confidence intervals.}

\item{round}{a scalar.}

\item{brackets}{a vector with character strings to replace left and right brackets.}

\item{collapse}{a string. separator for numbers within brackets}
}
\description{
glue_bracket
}
