% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{hash_replace}
\alias{hash_replace}
\title{hash_replace}
\usage{
hash_replace(string, sample = NULL, envir = parent.frame())
}
\arguments{
\item{string}{text to search in}

\item{sample}{default NULL. If provided, samples a vector}

\item{envir}{the environment to call replacements from. Defaults to parent frame}
}
\description{
Finds are replaces hash codes, replaces with objects of the same name
}
