% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{initiate_targets}
\alias{initiate_targets}
\title{initiate_targets}
\usage{
initiate_targets(path)
}
\arguments{
\item{path}{path to targets directory}
}
\description{
Creates a boilerplate targets directory
}
