% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_paper.R
\name{introduction}
\alias{introduction}
\title{introduction}
\usage{
introduction(known = "", unknown = "", objectives = "")
}
\arguments{
\item{known}{What is currently known about the topic}

\item{unknown}{what is unknown?}

\item{objectives}{what are the objectives of the current study?}
}
\description{
Scaffold an introduction
}
