% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{m_sd}
\alias{m_sd}
\title{m_sd}
\usage{
m_sd(x, na.rm = TRUE, digits = 2, pattern = "m (SD = {sd})")
}
\arguments{
\item{x}{vector}

\item{na.rm}{logical, should missing values be removed?}

\item{digits}{number of digits to round to}

\item{pattern}{glue pattern}
}
\description{
Print mean and sd
}
