% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_model_formats.R
\name{mm_table}
\alias{mm_table}
\title{mm_table}
\usage{
mm_table(
  model,
  transf = NULL,
  transf_name = NULL,
  round = 2,
  round_p = 3,
  fixed_names = NULL,
  simple_names = F,
  collapse = " - ",
  brackets = c("(", ")")
)
}
\arguments{
\item{model}{a model of glmerMod}

\item{transf}{function to transform estimates}

\item{transf_name}{string, used to rename 95\% CI column}

\item{round}{a scalar, defaults to 2}

\item{round_p}{a scalar. The number of digits to round p to.}

\item{fixed_names}{a vector of predictor names}

\item{simple_names}{a bool. If True, simple names are given}

\item{collapse}{a string. Value to separate confidence intervals with}

\item{brackets}{a vector. passed to glue, bracket.}
}
\description{
Creates a table of fixed and random effects
}
