% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{most_freq}
\alias{most_freq}
\title{most_freq}
\usage{
most_freq(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{bool. Should missing values be removed?}
}
\description{
Returns most frequent value
}
