% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{open_citation}
\alias{open_citation}
\title{open_citation}
\usage{
open_citation(package = NULL)
}
\arguments{
\item{package}{the package name}
}
\description{
Write a bibtex citation to a temporary file and open it. This may assist with loading R references into Zotero
}
