% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papaja_support.R
\name{papaja_affiliation}
\alias{papaja_affiliation}
\title{papaja_affiliation}
\usage{
papaja_affiliation(
  researchers,
  affiliations,
  corresponding,
  corres_address,
  corres_email,
  return_clipboard = TRUE
)
}
\arguments{
\item{researchers}{list, name, reference}

\item{affiliations}{list, reference, affiliation string}

\item{corresponding}{character, name (must match researcher name perfectly)}

\item{corres_address}{character}

\item{corres_email}{character}

\item{return_clipboard}{if TRUE results are not printer but sent to the clipboard}
}
\description{
Papaja affiliations
}
\examples{
library(papyr)
papaja_affiliation(
 researchers = list(
   "James H. Conigrave" = c("aff1"),
   "Other person" = c("aff3", "aff2")
 ),
 affiliations = list(
   "aff1" = "Affiliation 1",
   "aff2" = "Affiliation 2",
   "aff3" = "Affiliation 3"
 ),
 corresponding = "James H. Conigrave",
 corres_address = "Drug Health Services, Level 6, King George V Building, Royal Prince Alfred Hospital, Camperdown NSW 2050, AUSTRALIA",
 corres_email = "gmail.com",
 return_clipboard = FALSE
)
}
