% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pie_chart}
\alias{pie_chart}
\title{pie_chart}
\usage{
pie_chart(vector, colours = NULL)
}
\arguments{
\item{vector}{a vector}

\item{colours}{a vector of hex codes.}
}
\description{
Rounds p value to specified digits and uses less symbol if result it zero.
}
