% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{prevent_duplicates}
\alias{prevent_duplicates}
\title{prevent_duplicates}
\usage{
prevent_duplicates(path)
}
\arguments{
\item{path}{character string to path}
}
\description{
Stops users accidentally overriding important files
}
