% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{print_t}
\alias{print_t}
\title{print_t}
\usage{
print_t(x, round = 2, pattern = "$t$ = {t}, $df$ = {df}; $p$ = {p}")
}
\arguments{
\item{x}{the t.test --- an object of class "htest"}
}
\description{
print the contents of a t.test in APA format
}
