% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{q_alpha}
\alias{q_alpha}
\title{q_alpha}
\usage{
q_alpha(x, round = 2)
}
\arguments{
\item{x}{a dataframe}

\item{round}{a numeric. Result will be rounded to this number.}
}
\description{
Calculates and rounds alphas
}
