% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{raincloud}
\alias{raincloud}
\title{raincloud}
\usage{
raincloud(data, ..., palette = "Spectral", transition = NULL)
}
\arguments{
\item{data}{data}

\item{...}{variables}

\item{palette}{String of palette defining ggplot2 colours}

\item{transition}{variable to transition over}
}
\description{
raincloud
}
