% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif_change.R
\name{run_if_change}
\alias{run_if_change}
\title{run_if_change}
\usage{
run_if_change(
  code = NULL,
  object.names = NULL,
  compare = NULL,
  path = NULL,
  run = "auto"
)
}
\arguments{
\item{code}{a code chunk wrapped in\ \{\}}

\item{object.names}{a character vector. The names to be saved and loaded}

\item{compare}{an object. The object which if changed, will cause the code to run.}

\item{path}{a string. The path to the tempory file to be used. Must end in '.rData'}

\item{run}{a bool. If True, the analyses will be re-run even if the comparitor is identical.}
}
\description{
Takes in R commands and the path to a location to store temporary files. The function will create the file if it doesn't already exist, store the objects there, and only re-run if there has been a change to a reference object.
}
\details{
Runs analyses and then caches them. Analyses are returned in a list.
}
