% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdown.R
\name{safe_trackdown_upload}
\alias{safe_trackdown_upload}
\title{safe_trackdown_upload}
\usage{
safe_trackdown_upload(file, ...)
}
\arguments{
\item{file}{file path to rmd to upload}

\item{...}{other argumetns passed to trackdown::upload_file}
}
\description{
safe_trackdown_upload
}
