% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{to_docx}
\alias{to_docx}
\title{to_docx}
\usage{
to_docx(
  table,
  path,
  title = NULL,
  note = NULL,
  landscape = F,
  save_over = T,
  ...
)
}
\arguments{
\item{table}{a dataframe or list of data.frames}

\item{path}{a string. Where you want to save the file.}

\item{title}{string. If provided, allows tables to be named}

\item{note}{a string. Allows for notes}

\item{landscape}{a bool. If true, outputs a landscape word doc}

\item{save_over}{a bool. If true, to_docx will save over files with same path}

\item{...}{additional arguements passed to apa_table}
}
\description{
Sends a data.frame to a word doc.
}
