% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{to_pdf}
\alias{to_pdf}
\title{to_pdf}
\usage{
to_pdf(
  table,
  path,
  title = NULL,
  note = NULL,
  landscape = F,
  save_over = T,
  ...
)
}
\arguments{
\item{table}{a dataframe or list of data.frames. Alternatively, a list of LaTeX table code in character format If table == "clipboard" then copied latex code will be saved into the docx.}

\item{path}{a string. Where you want to save the file.}

\item{title}{string. If provided, allows tables will be named}

\item{note}{a string. Allows for notes}

\item{landscape}{a bool. If true, outputs a landscape word doc}

\item{save_over}{a bool. If true, to_docx will save over files with same path}

\item{...}{additional arguements passed to apa_table}
}
\description{
Sends a data.frame to a pdf doc.
}
