% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papaja_assist.R
\name{to_rowhead}
\alias{to_rowhead}
\title{to_rowhead}
\usage{
to_rowhead(data, x, italics = FALSE)
}
\arguments{
\item{data}{the table}

\item{x}{the columnt to convert to a rowheader}

\item{italics}{should row headers be italicized?}
}
\description{
Takes a frist column of a table and converts to a rowheader
}
