% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{word_apa}
\alias{word_apa}
\title{word_apa}
\usage{
word_apa(...)
}
\arguments{
\item{...}{arguments passed  to papaja::apa_table}
}
\description{
A wrapper for apa_table which will work nicely for word
}
