% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{zotero_notes}
\alias{zotero_notes}
\title{zotero_notes}
\usage{
zotero_notes(
  csv,
  path,
  title = "Zotero notes",
  date = format(Sys.time(), "\%d \%B, \%Y")
)
}
\arguments{
\item{csv}{an exported csv from zotero. Study type stored in the extra field}

\item{path}{a string. must end in .html}

\item{title}{a string. Defaults to "Zotero notes"}

\item{date}{a string. Defaults to current date}
}
\description{
Sends a data.frame to a word doc.
}
