% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{m_iqr}
\alias{m_iqr}
\title{m_iqr}
\usage{
m_iqr(x, round = 1, quantiles = F, na.rm = T, in.brack = F)
}
\arguments{
\item{x}{a character vector}

\item{round}{a numeric.}

\item{quantiles}{a bool.}

\item{na.rm}{a bool.}

\item{in.brack}{a bool.}
}
\description{
takes in a numeric vector, calculates a median and iqr and returns as text.
}
