% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_word.R
\name{n_percent}
\alias{n_percent}
\title{n_percent}
\usage{
n_percent(logical, round = 2, na.rm = T, pattern = "{n} ({p}\%)")
}
\arguments{
\item{logical}{a logical vector to count}

\item{round}{scalar, the number of digits}

\item{na.rm}{bool, whether to remove NAs}

\item{pattern}{a character string using glue syntax. Variable names are 'n' and 'p'.}
}
\value{
a character
}
\description{
Returns the number than meet a condition, and the percentage in brackets
}
